Feature/refactor pipeline to new dir (#557)
* up

* up
