--- r2r/base/abstractions/error.py
@@ -1,10 +1,0 @@
-from typing import Any, Optional
-
-
-class R2RException(Exception):
-    def __init__(
-        self, message: str, status_code: int, detail: Optional[Any] = None
-    ):
-        self.message = message
-        self.status_code = status_code
-        super().__init__(self.message)
