--- r2r/base/abstractions/user.py
@@ -1,11 +1,0 @@
-import uuid
-from typing import List
-
-from pydantic import BaseModel
-
-
-class UserStats(BaseModel):
-    user_id: uuid.UUID
-    num_files: int
-    total_size_in_bytes: int
-    document_ids: List[uuid.UUID]
