+++ r2r/main/api/routes/base_router.py
@@ -3,7 +3,8 @@ import logging
 
 from fastapi import APIRouter, HTTPException
 
-from r2r.base import R2RException, manage_run
+from r2r.base import manage_run
+from r2r.main.abstractions import R2RException
 
 logger = logging.getLogger(__name__)
