+++ r2r/main/services/retrieval_service.py
@@ -5,13 +5,13 @@ from typing import Optional
 
 from r2r.base import (
     KVLoggingSingleton,
-    R2RException,
     RunManager,
     manage_run,
     to_async_generator,
 )
 from r2r.base.abstractions.llm import GenerationConfig
 from r2r.base.abstractions.search import KGSearchSettings, VectorSearchSettings
+from r2r.main.abstractions import R2RException
 from r2r.pipes import EvalPipe
 from r2r.telemetry.telemetry_decorator import telemetry_event
